Fixed compilation problem under windows

+++ tpd_DB/ps_out.cpp
@@ -198,8 +198,8 @@ void PSFile::pspage_header(const DBbox box)
 {
    double W=((220.0 - 40.0)/25.4)*72.0;
    double H=((297.0 - 40.0)/25.4)*72.0;
-   double w = fabs(box.p1().x() - box.p2().x());
-   double h = fabs(box.p1().y() - box.p2().y());
+   double w = fabs(double(box.p1().x() - box.p2().x()));
+   double h = fabs(double(box.p1().y() - box.p2().y()));
    double sc = (W/H < w/h) ? w/W : h/H;
    double tx = ((box.p1().x() + box.p2().x()) - ( W   * sc) ) / 2;
    double ty = ((box.p1().y() + box.p2().y()) - ( H   * sc) ) / 2;
